+++ src/Components/Contexts.tsx
@@ -7,7 +7,7 @@ import { defaultColorKey } from "../constants";
 export const DatabaseContext = createContext<Database>({});
 export const DataPathsContext = createContext<DataPaths>({});
 export const AddAlertContext = createContext<AddAlert>(() => {});
-export const ColorKeyContext = createContext<() => ColorKey>(
+export const ColorKeyContext = createContext<(colorKey: ColorKey) => void>(
   () => defaultColorKey
 );
 export const UserContext = createContext<User>(undefined);
